backend/requirements.txt

# FastAPI and web framework
fastapi
uvicorn
python-multipart

# Database
sqlalchemy
psycopg2-binary
alembic

# Authentication and Security
python-jose
passlib[bcrypt]

# Validation and Configuration
pydantic
pydantic-settings
email-validator

# AWS SDK for S3
boto3
botocore

# Additional utilities
python-magic


# File handling and utilities
python-dotenv
